+++ app/src/main/kotlin/org/koitharu/kotatsu/list/ui/adapter/MangaListDetailedItemAD.kt
@@ -37,5 +37,6 @@ fun mangaListDetailedItemAD(
 		binding.textViewTags.text = item.tags.joinToString(separator = ", ") { it.title ?: "" }
 		binding.badge.number = item.counter
 		binding.badge.isVisible = item.counter > 0
+		binding.textViewBroken.isVisible = item.isBroken
 	}
 }