+++ app/src/main/kotlin/org/koitharu/kotatsu/list/ui/adapter/MangaListItemAD.kt
@@ -26,5 +26,6 @@ fun mangaListItemAD(
 		binding.imageViewCover.setImageAsync(item.coverUrl, item.manga)
 		binding.badge.number = item.counter
 		binding.badge.isVisible = item.counter > 0
+		binding.textViewBroken.isVisible = item.isBroken
 	}
 }